+++ app/src/ui/changes/changed-file.tsx
@@ -1,51 +1,22 @@
 import * as React from 'react'
-import * as Path from 'path'
 
 import { AppFileStatus, mapStatus, iconForStatus } from '../../models/status'
 import { PathLabel } from '../lib/path-label'
 import { Octicon } from '../octicons'
-import { showContextualMenu } from '../main-process-proxy'
 import { Checkbox, CheckboxValue } from '../lib/checkbox'
-import { IMenuItem } from '../../lib/menu-item'
-
-const GitIgnoreFileName = '.gitignore'
-
-const RestrictedFileExtensions = ['.cmd', '.exe', '.bat', '.sh']
-
-const defaultEditorLabel = __DARWIN__
-  ? 'Open in External Editor'
-  : 'Open in external editor'
 
 interface IChangedFileProps {
   readonly path: string
   readonly status: AppFileStatus
   readonly oldPath?: string
   readonly include: boolean | null
-  readonly onIncludeChanged: (path: string, include: boolean) => void
-  readonly onDiscardChanges: (path: string) => void
-  readonly onDiscardAllChanges: () => void
-  /**
-   * Called to reveal a file in the native file manager.
-   * @param path The path of the file relative to the root of the repository
-   */
-  readonly onRevealInFileManager: (path: string) => void
-
-  /**
-   * Called to open a file it its default application
-   * @param path The path of the file relative to the root of the repository
-   */
-  readonly onOpenItem: (path: string) => void
   readonly availableWidth: number
-  readonly onIgnore: (pattern: string) => void
-
-  /** The name of the currently selected external editor */
-  readonly externalEditorLabel?: string
-
-  /**
-   * Called to open a file using the user's configured external editor
-   * @param path The path of the file relative to the root of the repository
-   */
-  readonly onOpenInExternalEditor: (path: string) => void
+  readonly onIncludeChanged: (path: string, include: boolean) => void
+  readonly onContextMenu: (
+    path: string,
+    status: AppFileStatus,
+    event: React.MouseEvent<any>
+  ) => void
 }
 
 /** a changed file in the working directory for a given repository */
@@ -108,73 +79,7 @@ export class ChangedFile extends React.Component<IChangedFileProps, {}> {
     )
   }
 
-  private onContextMenu = (event: React.MouseEvent<any>) => {
-    event.preventDefault()
-
-    const extension = Path.extname(this.props.path)
-    const fileName = Path.basename(this.props.path)
-    const isSafeExtension = __WIN32__
-      ? RestrictedFileExtensions.indexOf(extension.toLowerCase()) === -1
-      : true
-    const revealInFileManagerLabel = __DARWIN__
-      ? 'Reveal in Finder'
-      : __WIN32__ ? 'Show in Explorer' : 'Show in your File Manager'
-    const openInExternalEditor = this.props.externalEditorLabel
-      ? `Open in ${this.props.externalEditorLabel}`
-      : defaultEditorLabel
-
-    const items: IMenuItem[] = [
-      {
-        label: __DARWIN__ ? 'Discard Changes…' : 'Discard changes…',
-        action: () => this.props.onDiscardChanges(this.props.path),
-      },
-      {
-        label: __DARWIN__ ? 'Discard All Changes…' : 'Discard all changes…',
-        action: () => this.props.onDiscardAllChanges(),
-      },
-      { type: 'separator' },
-      {
-        label: 'Ignore',
-        action: () => this.props.onIgnore(this.props.path),
-        enabled: fileName !== GitIgnoreFileName,
-      },
-    ]
-
-    if (extension.length) {
-      items.push({
-        label: __DARWIN__
-          ? `Ignore All ${extension} Files`
-          : `Ignore all ${extension} files`,
-        action: () => this.props.onIgnore(`*${extension}`),
-        enabled: fileName !== GitIgnoreFileName,
-      })
-    }
-
-    items.push(
-      { type: 'separator' },
-      {
-        label: revealInFileManagerLabel,
-        action: () => this.props.onRevealInFileManager(this.props.path),
-        enabled: this.props.status !== AppFileStatus.Deleted,
-      },
-      {
-        label: openInExternalEditor,
-        action: this.openInExternalEditor,
-        enabled: isSafeExtension && this.props.status !== AppFileStatus.Deleted,
-      },
-      {
-        label: __DARWIN__
-          ? 'Open with Default Program'
-          : 'Open with default program',
-        action: () => this.props.onOpenItem(this.props.path),
-        enabled: isSafeExtension && this.props.status !== AppFileStatus.Deleted,
-      }
-    )
-
-    showContextualMenu(items)
-  }
-
-  private openInExternalEditor = () => {
-    this.props.onOpenInExternalEditor(this.props.path)
+  private onContextMenu = (event: React.MouseEvent<HTMLDivElement>) => {
+    this.props.onContextMenu(this.props.path, this.props.status, event)
   }
 }
