+++ app/src/ui/changes/changes-list.tsx
@@ -1,11 +1,13 @@
 import * as React from 'react'
+import * as Path from 'path'
+
 import { CommitMessage } from './commit-message'
 import { ChangedFile } from './changed-file'
 import { List, ClickSource } from '../lib/list'
-
 import {
   WorkingDirectoryStatus,
   WorkingDirectoryFileChange,
+  AppFileStatus,
 } from '../../models/status'
 import { DiffSelectionType } from '../../models/diff'
 import { CommitIdentity } from '../../models/commit-identity'
@@ -21,6 +23,11 @@ import { ITrailer } from '../../lib/git/interpret-trailers'
 import { IMenuItem } from '../../lib/menu-item'
 
 const RowHeight = 29
+const RestrictedFileExtensions = ['.cmd', '.exe', '.bat', '.sh']
+const defaultEditorLabel = __DARWIN__
+  ? 'Open in External Editor'
+  : 'Open in external editor'
+const GitIgnoreFileName = '.gitignore'
 
 interface IChangesListProps {
   readonly repository: Repository
@@ -120,14 +127,8 @@ export class ChangesList extends React.Component<IChangesListProps, {}> {
         include={includeAll}
         key={file.id}
         onIncludeChanged={this.props.onIncludeChanged}
-        onDiscardChanges={this.onDiscardChanges}
-        onDiscardAllChanges={this.onDiscardAllChanges}
-        onRevealInFileManager={this.props.onRevealInFileManager}
-        onOpenItem={this.props.onOpenItem}
         availableWidth={this.props.availableWidth}
-        onIgnore={this.props.onIgnore}
-        externalEditorLabel={this.props.externalEditorLabel}
-        onOpenInExternalEditor={this.props.onOpenInExternalEditor}
+        onContextMenu={this.onItemContextMenu}
       />
     )
   }
@@ -171,6 +172,75 @@ export class ChangesList extends React.Component<IChangesListProps, {}> {
     showContextualMenu(items)
   }
 
+  private onItemContextMenu = (
+    path: string,
+    status: AppFileStatus,
+    event: React.MouseEvent<any>
+  ) => {
+    event.preventDefault()
+
+    const extension = Path.extname(path)
+    const fileName = Path.basename(path)
+    const isSafeExtension = __WIN32__
+      ? RestrictedFileExtensions.indexOf(extension.toLowerCase()) === -1
+      : true
+    const revealInFileManagerLabel = __DARWIN__
+      ? 'Reveal in Finder'
+      : __WIN32__ ? 'Show in Explorer' : 'Show in your File Manager'
+    const openInExternalEditor = this.props.externalEditorLabel
+      ? `Open in ${this.props.externalEditorLabel}`
+      : defaultEditorLabel
+    const items: IMenuItem[] = [
+      {
+        label: __DARWIN__ ? 'Discard Changes…' : 'Discard changes…',
+        action: () => this.onDiscardChanges(path),
+      },
+      {
+        label: __DARWIN__ ? 'Discard All Changes…' : 'Discard all changes…',
+        action: () => this.onDiscardAllChanges(),
+      },
+      { type: 'separator' },
+      {
+        label: 'Ignore',
+        action: () => this.props.onIgnore(path),
+        enabled: fileName !== GitIgnoreFileName,
+      },
+    ]
+
+    if (extension.length) {
+      items.push({
+        label: __DARWIN__
+          ? `Ignore All ${extension} Files`
+          : `Ignore all ${extension} files`,
+        action: () => this.props.onIgnore(`*${extension}`),
+        enabled: fileName !== GitIgnoreFileName,
+      })
+    }
+
+    items.push(
+      { type: 'separator' },
+      {
+        label: revealInFileManagerLabel,
+        action: () => this.props.onRevealInFileManager(path),
+        enabled: status !== AppFileStatus.Deleted,
+      },
+      {
+        label: openInExternalEditor,
+        action: () => this.props.onOpenInExternalEditor(path),
+        enabled: isSafeExtension && status !== AppFileStatus.Deleted,
+      },
+      {
+        label: __DARWIN__
+          ? 'Open with Default Program'
+          : 'Open with default program',
+        action: () => this.props.onOpenItem(path),
+        enabled: isSafeExtension && status !== AppFileStatus.Deleted,
+      }
+    )
+
+    showContextualMenu(items)
+  }
+
   public render() {
     const fileList = this.props.workingDirectory.files
     const selectedRow = fileList.findIndex(
@@ -181,10 +251,6 @@ export class ChangesList extends React.Component<IChangesListProps, {}> {
     const filesDescription = `${fileCount} changed ${filesPlural}`
     const anyFilesSelected =
       fileCount > 0 && this.includeAllValue !== CheckboxValue.Off
-    const invalidationProps = {
-      workingDirectory: this.props.workingDirectory,
-      externalEditorLabel: this.props.externalEditorLabel,
-    }
 
     return (
       <div className="changes-list-container file-list">
@@ -204,7 +270,7 @@ export class ChangesList extends React.Component<IChangesListProps, {}> {
           rowRenderer={this.renderRow}
           selectedRow={selectedRow}
           onSelectionChanged={this.props.onFileSelectionChanged}
-          invalidationProps={invalidationProps}
+          invalidationProps={this.props.workingDirectory}
           onRowClick={this.props.onRowClick}
         />
